+++ src/proxy.js
@@ -1,21 +1,23 @@
+// @flow
 /* eslint no-param-reassign: 0 */
 import ValidationError from './error';
 import { checkRules, applySanitizers } from './validation';
+import type { ProxyFactory } from './types';
 
-export default function ProxyValidator(schema, sanitizers) {
+export default function ProxyValidator(schema: Object, sanitizersSchema: Object): ProxyFactory {
   const handler = {
-    set(object, prop, value) {
+    set(object: Object, prop: string, value: string): boolean {
       const { [prop]: rules } = schema;
       const { success, errors } = checkRules(rules, value);
       if (success) {
-        const { [prop]: transforms } = sanitizers;
-        object[prop] = applySanitizers(transforms, value);
+        const { [prop]: sanitizers } = sanitizersSchema;
+        object[prop] = applySanitizers(sanitizers, value);
         return success;
       }
-      throw new ValidationError(errors);
+      throw new ValidationError({ [prop]: errors });
     }
   };
-  return function Proxied() {
+  return function Proxied(): Object {
     const target = {};
     return new Proxy(target, handler);
   };